practic_p/pyproject.toml
[tool.poetry]
name = "practic-p"
version = "0.1.0"
description = ""
authors = ["Pi4kovaMasha <pi4kova.masha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
uvicorn = "^0.32.1"
pydantic-settings = "^2.6.1"
psycopg2 = "^2.9.10"
asyncio = "^3.4.3"
fastapi = "^0.115.5"
alembic = "^1.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
